fixed: organization data not shown to anon user
- Added default public fields inside the loader

This feels weird because the actual data (which fields are public)
is not represented in our model. We extend the model data inside
profile/$username and now also organization/$slug.

In my opinion the fields that are always public (f.e. logo, etc...)
should also be present in the database publicFields.