Reworked checkIdentityOrThrow()
Removed multipartBoolean (not needed anymore as we found another solution for the thing it had done).
Renamed userId to formSenderId to clarify what we are checking.
We are checking that the form sender is also the session user.
But he always is, isn't he? So i'm a bit confused if we need this method (see issue "Do we need checkIdentityOrThrow()?")