Rename areaNames to areas to support filtering
On profile explore page we got a raw query which selects the area names as string array in a field on the profile named areaNames.
This field is not implemented in the filter function. Thats why i renamed the field to areas, which is present in the filter function.
I did the same on organization explore page with the fields areaNames and organizationTypeTitles. They are both selected as areas and types now.
It also occured on event detail page where we use a raw query to get the full depth speakers/participants/... There i changed the field eventName to participatedEvents