Updated all packages and fixed almost all typescript compilation errors
Two compilation errors could not yet be fixed:
- @remix-run/web-fetch still throws a compilation error concerning the Headers type (Known issue to remix developers: https://github.com/remix-run/remix/issues/7067)
- @testing-library/cypress/types throws a compilation error. Did not yet look into it but may be caused by conflicting globals of cypress and jest.

Apart of the compilation errors the app throws an error on build time concerning the use of commonJS syntax (SyntaxError: Cannot use import statement outside a module). Thats a tricky one and ive not found a solution yet.
Resources:
- https://github.com/remix-run/remix/issues/4234
- https://github.com/remix-run/remix/discussions/4309
- https://community.fly.io/t/cannot-use-import-statement-outside-a-module/13533
- https://stackoverflow.com/questions/58384179/syntaxerror-cannot-use-import-statement-outside-a-module (Maybe unrelated)
- https://remix.run/docs/en/main/guides/gotchas (Maybe unrelated)

Last but not least, the tests are not running, too. Did not yet look into it deeply but it seems it also has something to do with the globals of cypress and jest that overlap. The error Messages point to jest-setup.ts installGlobals() call.