Optimize i18n locales serving
Needed to flatten out the locales folder structure into one json for now to avoid following two alternatives:
- dynamically creating a resource object in i18n.ts which is cost intensive. (see last commits)
- Hardcoding the resource object structure with all namespaces in full depth and importing each namespace .json file to add it to this object structure (Maybe we can do this programmatically but i stopped here for the moment)

If we want to revert back to folder structure with multiple jsons, we need to import every single json in i18n.ts so that the backend route (routes/i18n/locales.ts) can access them. Originally the problem occured due to the reason, that the namespaces with the correct i18n setup get requested via searchParams and "/"'s messed up those search params.

Maybe there is a better solution to this and i overlooked it, but for now it works well.

- It only fetches the required namespaces and locales for the specific route your on, which is what was intended.
- Also the new setup avoids the flickering we saw on the earlier setup because the server now holds the truth. - Language settings are still persisted via cookie
- Locales still get cached (on production) via the Cache-Control headers (Maybe we can use a more controlled cache in the next iteration f.e. with clientLoaders and localForage -> see https://www.youtube.com/watch?v=fuwm0BBNkRY)
- Works like a charm even without javascript enabled (earlier setup did not support the language switch without js)