Outsource components from __components into their own files
One file for each composed component. With that we reduce the client bundle size and as a side effect we have a better overview of which components already exist (twice or sometimes mor than twice...)

The Modal component which was located inside routes/__components.tsx is a perfect example:

Inside this module we exported <Modal.Root> which is used by root. So every route, even when they do not use a Modal, needs to fetch the whole module. Now ModalRoot is its own module and every route has to fetch only this small module. Only if a route really uses a <Modal> it gets fetched.

You can apply this example to every __components module, that defines components, that are used in different routes. All routes that use this module have to fetch everything even if they dont need it.